fix: router feat: add modules feat: change config -> modules
